Clean up Tabs animation logic (#65878)
* Clean up and misc improvements.

* Move utility from ToggleGroupControl to utils.

* Clean up Tabs.Tablist

* Clean up ToggleGroupControl and update utility from latest trunk version

* Clean up scroll utility

* Remove unnecessary null.

* CHANGELOG

---------
